automatically link/unlink the agent on machine start/stop
At @GSA, we were running into the issue of Nessus Agents remaining
linked even after their virtual machine was gone. Rather than having
Ansible do the linking, this instead uses a systemd service to link and
unlink the Agent when the service (and thus the machine) is started or
stopped, respectively.